ref: remove EngineDecorator and extend base engine
- use multiple shards to books index
- sort by db_id in tests
- override base engine's update and delete to implement ShardRouting logic
